speech_to_text_google/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech_to_text_google"
version = "0.1.0"
description = "Google Cloud Speech-to-Text helper"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "google-cloud-speech>=2.19.0",
  "pydub>=0.25.1",
]
